# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cooperative Services API",
    "description" : "Runnable prototype. Development OTP and simulated payments are explicitly gated. See docs/api-contract.md for response examples and integration rules.",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/jobs/{id}/rating" : {
      "post" : {
        "tags" : [ "billing-controller" ],
        "operationId" : "rate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Rating"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/jobs/{id}/payments/simulate" : {
      "post" : {
        "tags" : [ "billing-controller" ],
        "operationId" : "pay",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/jobs/{id}/payment" : {
      "get" : {
        "tags" : [ "billing-controller" ],
        "operationId" : "payment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/jobs/{id}/invoice" : {
      "get" : {
        "tags" : [ "billing-controller" ],
        "operationId" : "invoice",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/catalog/subservices" : {
      "get" : {
        "tags" : [ "catalog-controller" ],
        "operationId" : "services",
        "parameters" : [ {
          "name" : "categoryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/welfare" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "welfare",
        "parameters" : [ {
          "name" : "societyId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Rating" : {
        "type" : "object",
        "properties" : {
          "stars" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 5,
            "minimum" : 1
          },
          "feedback" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}